agent_youtuber/requirements.txt
crewai>=0.102.0
python-dotenv>=1.0.0
litellm>=1.15.0
langchain>=0.1.0
google-generativeai>=0.3.0 